tests/test_postgres_functions.py
from asyncpg import Connection
from random import Random
from testcontainers_demo.postgres_functions import insert_random_key_value_pair, get_value, get_count
import pyperclip

async def test_postgres(test_database: Connection):
    rng = Random(0)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


# this will generate a ton of unit tests
async def test_generate_tests():
    test_cases = []

    for i in range(0, 100):
        s = f'''
async def test_postgres_{i}(test_database: Connection):
    rng = Random({i})
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 
'''
        test_cases.append(s)

    full_test_script = "\n".join(test_cases)
    pyperclip.copy(full_test_script)
    print("All test cases have been copied to the clipboard.")


async def test_postgres_0(test_database: Connection):
    rng = Random(0)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_1(test_database: Connection):
    rng = Random(1)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_2(test_database: Connection):
    rng = Random(2)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_3(test_database: Connection):
    rng = Random(3)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_4(test_database: Connection):
    rng = Random(4)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_5(test_database: Connection):
    rng = Random(5)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_6(test_database: Connection):
    rng = Random(6)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_7(test_database: Connection):
    rng = Random(7)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_8(test_database: Connection):
    rng = Random(8)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_9(test_database: Connection):
    rng = Random(9)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_10(test_database: Connection):
    rng = Random(10)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_11(test_database: Connection):
    rng = Random(11)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_12(test_database: Connection):
    rng = Random(12)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_13(test_database: Connection):
    rng = Random(13)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_14(test_database: Connection):
    rng = Random(14)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_15(test_database: Connection):
    rng = Random(15)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_16(test_database: Connection):
    rng = Random(16)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_17(test_database: Connection):
    rng = Random(17)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_18(test_database: Connection):
    rng = Random(18)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_19(test_database: Connection):
    rng = Random(19)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_20(test_database: Connection):
    rng = Random(20)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_21(test_database: Connection):
    rng = Random(21)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_22(test_database: Connection):
    rng = Random(22)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_23(test_database: Connection):
    rng = Random(23)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_24(test_database: Connection):
    rng = Random(24)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_25(test_database: Connection):
    rng = Random(25)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_26(test_database: Connection):
    rng = Random(26)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_27(test_database: Connection):
    rng = Random(27)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_28(test_database: Connection):
    rng = Random(28)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_29(test_database: Connection):
    rng = Random(29)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_30(test_database: Connection):
    rng = Random(30)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_31(test_database: Connection):
    rng = Random(31)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_32(test_database: Connection):
    rng = Random(32)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_33(test_database: Connection):
    rng = Random(33)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_34(test_database: Connection):
    rng = Random(34)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_35(test_database: Connection):
    rng = Random(35)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_36(test_database: Connection):
    rng = Random(36)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_37(test_database: Connection):
    rng = Random(37)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_38(test_database: Connection):
    rng = Random(38)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_39(test_database: Connection):
    rng = Random(39)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_40(test_database: Connection):
    rng = Random(40)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_41(test_database: Connection):
    rng = Random(41)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_42(test_database: Connection):
    rng = Random(42)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_43(test_database: Connection):
    rng = Random(43)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_44(test_database: Connection):
    rng = Random(44)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_45(test_database: Connection):
    rng = Random(45)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_46(test_database: Connection):
    rng = Random(46)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_47(test_database: Connection):
    rng = Random(47)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_48(test_database: Connection):
    rng = Random(48)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_49(test_database: Connection):
    rng = Random(49)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_50(test_database: Connection):
    rng = Random(50)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_51(test_database: Connection):
    rng = Random(51)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_52(test_database: Connection):
    rng = Random(52)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_53(test_database: Connection):
    rng = Random(53)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_54(test_database: Connection):
    rng = Random(54)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_55(test_database: Connection):
    rng = Random(55)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_56(test_database: Connection):
    rng = Random(56)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_57(test_database: Connection):
    rng = Random(57)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_58(test_database: Connection):
    rng = Random(58)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_59(test_database: Connection):
    rng = Random(59)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_60(test_database: Connection):
    rng = Random(60)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_61(test_database: Connection):
    rng = Random(61)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_62(test_database: Connection):
    rng = Random(62)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_63(test_database: Connection):
    rng = Random(63)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_64(test_database: Connection):
    rng = Random(64)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_65(test_database: Connection):
    rng = Random(65)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_66(test_database: Connection):
    rng = Random(66)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_67(test_database: Connection):
    rng = Random(67)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_68(test_database: Connection):
    rng = Random(68)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_69(test_database: Connection):
    rng = Random(69)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_70(test_database: Connection):
    rng = Random(70)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_71(test_database: Connection):
    rng = Random(71)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_72(test_database: Connection):
    rng = Random(72)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_73(test_database: Connection):
    rng = Random(73)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_74(test_database: Connection):
    rng = Random(74)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_75(test_database: Connection):
    rng = Random(75)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_76(test_database: Connection):
    rng = Random(76)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_77(test_database: Connection):
    rng = Random(77)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_78(test_database: Connection):
    rng = Random(78)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_79(test_database: Connection):
    rng = Random(79)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_80(test_database: Connection):
    rng = Random(80)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_81(test_database: Connection):
    rng = Random(81)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_82(test_database: Connection):
    rng = Random(82)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_83(test_database: Connection):
    rng = Random(83)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_84(test_database: Connection):
    rng = Random(84)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_85(test_database: Connection):
    rng = Random(85)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_86(test_database: Connection):
    rng = Random(86)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_87(test_database: Connection):
    rng = Random(87)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_88(test_database: Connection):
    rng = Random(88)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_89(test_database: Connection):
    rng = Random(89)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_90(test_database: Connection):
    rng = Random(90)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_91(test_database: Connection):
    rng = Random(91)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_92(test_database: Connection):
    rng = Random(92)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_93(test_database: Connection):
    rng = Random(93)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_94(test_database: Connection):
    rng = Random(94)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_95(test_database: Connection):
    rng = Random(95)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_96(test_database: Connection):
    rng = Random(96)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_97(test_database: Connection):
    rng = Random(97)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_98(test_database: Connection):
    rng = Random(98)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 


async def test_postgres_99(test_database: Connection):
    rng = Random(99)
    count = await get_count(test_database)
    assert count == 0
    k,v = await insert_random_key_value_pair(rng, test_database)
    retrieved = await get_value(k, test_database)
    assert v == retrieved
    count = await get_count(test_database)
    assert count == 1 
